Move TracingConfig to utils
