+++ compiler/src/compiler/hir_to_mir.rs
@@ -4,6 +4,7 @@ use super::{
     error::CompilerError,
     hir,
     mir::{Body, Expression, Id, Mir},
+    utils::TracingConfig,
 };
 use crate::{
     builtin_functions::BuiltinFunction,
@@ -17,23 +18,6 @@ use std::{collections::HashMap, sync::Arc};
 #[salsa::query_group(HirToMirStorage)]
 pub trait HirToMir: CstDb + AstToHir + LspPositionConversion {
     fn mir(&self, module: Module, config: TracingConfig) -> Option<Arc<Mir>>;
-}
-
-#[derive(PartialEq, Eq, Hash, Debug, Clone)]
-pub struct TracingConfig {
-    pub register_fuzzables: bool,
-    pub trace_calls: bool,
-    pub trace_evaluated_expressions: bool,
-}
-
-impl TracingConfig {
-    pub fn none() -> Self {
-        Self {
-            register_fuzzables: false,
-            trace_calls: false,
-            trace_evaluated_expressions: false,
-        }
-    }
 }
 
 fn mir(db: &dyn HirToMir, module: Module, config: TracingConfig) -> Option<Arc<Mir>> {
