+++ compiler/src/compiler/mod.rs
@@ -14,3 +14,5 @@ pub mod rcst;
 pub mod rcst_to_cst;
 pub mod string_to_rcst;
 mod utils;
+
+pub use utils::TracingConfig;
